+++ src/main/java/minestrapteam/minestrappolation/block/ore/BlockSoulOre.java
@@ -13,16 +13,18 @@ import net.minecraft.world.World;
 import net.minecraftforge.common.EnumPlantType;
 import net.minecraftforge.common.IPlantable;
 
-public class BlockSoulOre extends MBlockOre{
-
-	public BlockSoulOre(Material materialIn, MapColor mapColorIn, Item itemDrop, int meta, int expMin, int expMax, int dropAmount, int bonusAmount, String tool, int level, boolean silkHarvest) {
-		super(materialIn, mapColorIn, itemDrop, meta, expMin, expMax, dropAmount,bonusAmount, tool, level, silkHarvest);
+public class BlockSoulOre extends MBlockOre
+{
+	
+	public BlockSoulOre(Material materialIn, MapColor mapColorIn, Item itemDrop, int meta, int expMin, int expMax, int dropAmount, int bonusAmount, String tool, int level, boolean silkHarvest)
+	{
+		super(materialIn, mapColorIn, itemDrop, meta, expMin, expMax, dropAmount, bonusAmount, tool, level, silkHarvest);
 	}
-
+	
 	@Override
 	public AxisAlignedBB getCollisionBoundingBox(World worldIn, BlockPos pos, IBlockState state)
 	{
-		return new AxisAlignedBB((double)pos.getX(), (double)pos.getY(), (double)pos.getZ(), (double)pos.getX() + 1, (double)pos.getY() + 0.875F, (double)pos.getZ() + 1);
+		return new AxisAlignedBB(pos.getX(), pos.getY(), pos.getZ(), (double) pos.getX() + 1, (double) pos.getY() + 0.875F, (double) pos.getZ() + 1);
 	}
 	
 	@Override
